Simon/ot based ecdsa presign triples feature disable (#63)
Reimplements the separation according to the documentation .
The implementation and documentation both prevent running presign using two triples each coming from a different set of participants
Very important, the change in the API of presign now implies a change in the MPC API

+++ src/confidential_key_derivation/protocol.rs
@@ -112,7 +112,7 @@ pub fn ckd(
     // not enough participants
     if participants.len() < 2 {
         return Err(InitializationError::NotEnoughParticipants {
-            participants: participants.len() as u32,
+            participants: participants.len(),
         });
     };
